pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easy-fints"
version = "1.2.0"
description = "Python library and FastAPI wrapper around python-fints"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
  { name = "qhle" },
]
license-files = ["LICENSE"]
keywords = ["fints", "hbci", "fastapi", "banking", "rest", "api"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "fints",
  "Markdown",
  "python-dotenv",
]

[project.optional-dependencies]
dev = [
  "build",
  "pytest",
]
[project.scripts]
fints-rest-server = "easy_fints.cli:main"

[project.urls]
Homepage = "https://github.com/QHLe/easy-fints"
Repository = "https://github.com/QHLe/easy-fints"
Issues = "https://github.com/QHLe/easy-fints/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["easy_fints*"]
